// src/app/transcations/term-deposits-transcation/operations/interest-payment/shared/interest-payment.model.ts
export class InterestPayment {
    id: any;

    admissionNumber: any;

    accountNumber: any;

    termAccId: any;

    interestPaymentSummaryId: any;

    interestAmount: any;

    interestPostingDate: any;

    foreClosureInterestAmount: any;

    days: any;

    products: any;

    foreClosureProducts: any;

    status: any;

    statusName: any;

    transactionId: any;

    pacsId: any;

    pacsCode: any;

    remarks: any;

    transcationDate: any;

    transactionMode: any;

    signedCopyPath: any;

    fdNonCummulativeAccountsTransactionDTO: any;

    filesDTOList: any;
    
    isVerified: any;

    multipartFileListForDocument: any;

    interestPostingDateVal:any;
    
    multipartFileListsignedCopyPath:any;
}